Add dynamic info for changing quantity on dialog
